cloudrun/requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
